buqfreq: Use cpu_load_slope correctly (AndreiLux)
The load slope is the delta from the current load to the last cpu load
sample.
It's only positive if the current load exceeds the last load, so why is
it scaling
down a step if the load is actually higher? It makes no sense.

Change it to correctly scale down a step if the load slope exceeds the
load
slope size.